Minor username change, bugfix for username click
Added an "@" symbol in front of the username, fixed a bug in the function
that handles what happens when a username is clicked that was caused by
the introduced "@". Also, the forEach command wasn't playing nicely with
the streams.users[username] array/object anymore, so I went for a for
loop.